requirements.txt
pandas==1.1.5
numpy==1.19.4
matplotlib==3.3.3
plotly==4.14.1
scipy==1.5.4
openpyxl==3.0.5
PyYAML==5.3.1
